Update 31
Played round with form 1 visuals some more hope you like...

1. Setup tab control border to still look 3D style but used four new
thinner labels to simulate gold shading inside the border as opposed to
the while  (3D border is now set to off).
3. Also added 16 more gold labels to setup search combo \ numeric boxes
the same way.
4. This also corrected the nagging issue where the numeric up down
border was different to the combobox ones and was also always a little
smaller and didn't line up quite right with the comboboxes.
4. And moved skins picture box  to the left a little to make room for
mule info textbox to fit to the right.